Moved test data.
