+++ tests/fs_odm/fs-db-mapper.test.js
@@ -2,6 +2,7 @@ const fs = require('fs/promises');
 const path = require('path');
 
 const FsDbMapper = require('../../fs_odm/fs-db-mapper');
+const { testUsers } = require('./test-data');
 
 const testDbPath = path.join(__dirname, '..', '..', 'db', 'test-db.json');
 const testDbMapper = new FsDbMapper(testDbPath);
@@ -9,17 +10,10 @@ const testDbMapper = new FsDbMapper(testDbPath);
 jest.mock('fs/promises');
 
 describe('FsDbMapper:', () => {
-  fs.readFile = jest.fn(() => JSON.stringify(testUsers));
-  fs.writeFile = jest.fn();
-
-  const testUsers = [
-    {
-      id: 'd40ddf50-386e-4d6c-a10a-2c08d599ab19',
-      name: 'Software Engineering School',
-      email: 'software@engineering.school',
-      password: '$2a$08$qT85kN0AQU1mnl3fCyskzOfw/qxuQBRmzzIP1qwlB0QohZ0OghqSu',
-    },
-  ];
+  beforeAll(() => {
+    fs.readFile = jest.fn(() => JSON.stringify(testUsers));
+    fs.writeFile = jest.fn();
+  });
 
   test('should read data from fs db', async () => {
     const result = await testDbMapper.read();
