+++ tests/fs_odm/test-data.js
@@ -1,0 +1,10 @@
+const testUsers = [
+  {
+    id: 'd40ddf50-386e-4d6c-a10a-2c08d599ab19',
+    name: 'Software Engineering School',
+    email: 'software@engineering.school',
+    password: '$2a$08$qT85kN0AQU1mnl3fCyskzOfw/qxuQBRmzzIP1qwlB0QohZ0OghqSu',
+  },
+];
+
+module.exports = { testUsers };
